rename config.h to configuration.h
This is a preparation step for the next commit.
