+++ i3bar/include/configuration.h
